# commons-auth-api/src/doc/swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Handles authentication + authorization with registration-, invite and password-reset flows.",
    "version": "4.2.0",
    "title": "commons-auth api",
    "contact": {
      "name": "rocketbase.io",
      "url": "https://www.rocketbase.io",
      "email": "info@rocketbase.io"
    }
  },
...
  "paths": {
...
    "/auth/reset-password": {
      "put": {
        "tags": [
          "forgot-password"
        ],
        "summary": "rest password",
        "description": "perform password-change with given verification-token.",
        "operationId": "resetPasswordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "performPasswordReset",
            "description": "performPasswordReset",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PerformPasswordResetRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/user": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "find",
        "description": "page thew users-list with optional filter parameters.",
        "operationId": "findUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "description": "default 25 max is 200",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page",
            "in": "query",
            "description": "default is 1",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "username containing case-insensitive"
          },
          {
            "name": "firstName",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "firstName containing case-insensitive"
          },
          {
            "name": "lastName",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "lastName containing case-insensitive"
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "email containing case-insensitive"
          },
          {
            "name": "freetext",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "searches in all properties containing (case-insensitive)"
          },
          {
            "name": "keyValue",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "user needs to have key with value match. format: keyValue=KEY;VALUE"
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "default true - to search for disabled you need to add to query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageableResult-AppUserRead"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "create",
        "description": "create a new user threw admin-rights. normal flow is disabled.",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "create",
            "description": "create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppUserCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AppUserRead"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "patch",
        "description": "allows to update only parts of an user without sending the hole object.",
        "operationId": "patchUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "update",
            "description": "update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppUserUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AppUserRead"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "delete",
        "description": "remove a user from database.",
        "operationId": "deleteUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/user/{id}/password": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "reset password",
        "description": "admins can reset password of user",
        "operationId": "restPasswordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "resetPassword",
            "description": "reset password of a user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppUserResetPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AppUserRead"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "AppUserCreate": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "keyValues": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "Allowed key chars are a-Z, 0-9 and _-.# (max 50 chars)"
        },
        "admin": {
          "type": "boolean"
        },
        "enabled": {
          "type": "boolean"
        }
      },
      "title": "AppUserCreate"
    },
...
    "AppUserRead": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "keyValues": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "lastLogin": {
          "type": "string",
          "format": "date-time"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "enabled": {
          "type": "boolean"
        }
      },
      "title": "AppUserRead"
    },
...
    "AppUserUpdate": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "keyValues": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "will removed key that have value of null; will only add/replace new/existing key values; not mentioned key will still stay the same;"
        },
        "password": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "AppUserUpdate"
    },
...
    "PageableResult-AppUserRead": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppUserRead"
          }
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "PageableResult-AppUserRead"
    },
...
    "AppUserResetPassword": {
      "type": "object",
      "properties": {
        "resetPassword": {
          "type": "string"
        }
      },
      "title": "AppUserResetPassword"
    },
...
    "PerformPasswordResetRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "verification": {
          "type": "string"
        }
      },
      "title": "PerformPasswordResetRequest"
    },
...
  }
}